# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MotoHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      parameters:
        - name: type
          in: query
          required: false
          schema: { type: string, enum: [moto, peca, servico] }
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: brand
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: sellerId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Item" }
    post:
      operationId: createItem
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateItem" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
    patch:
      operationId: updateItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateItem" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Item" }
    delete:
      operationId: deleteItem
      tags: [items]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      parameters:
        - name: buyerId
          in: query
          required: false
          schema: { type: integer }
        - name: sellerId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      parameters:
        - name: itemId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }
    post:
      operationId: createReview
      tags: [reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReviewInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Coupon" }
    post:
      operationId: createCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCouponInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Coupon" }
  /coupons/{id}:
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ValidateCouponInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidateCouponResult" }
...
components:
  schemas:
...
    Item:
      type: object
      required: [id, type, category, title, price, image, description, location, sellerId, status, createdAt, premium, stock, ratingAvg, ratingCount, condition]
      properties:
        id: { type: integer }
        type: { type: string, enum: [moto, peca, servico] }
        category: { type: string }
        title: { type: string }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        condition: { type: string, enum: [novo, usado] }
        price: { type: number }
        year: { type: ["integer", "null"] }
        mileage: { type: ["integer", "null"] }
        engineSize: { type: ["integer", "null"] }
        image: { type: string }
        description: { type: string }
        location: { type: string }
        sellerId: { type: integer }
        status: { type: string, enum: [active, sold, pending] }
        createdAt: { type: string }
        premium: { type: boolean }
        stock: { type: integer }
        ratingAvg: { type: number }
        ratingCount: { type: integer }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        workingHours: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        fuelType: { type: ["string", "null"] }
        optionals: { type: ["string", "null"] }
        tradeInfo: { type: ["string", "null"] }
        extras: { type: ["string", "null"] }
    CreateItem:
      type: object
      required: [type, title, price, description, location, sellerId, image]
      properties:
        type: { type: string, enum: [moto, peca, servico] }
        category: { type: ["string", "null"] }
        title: { type: string, minLength: 3 }
        brand: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        condition: { type: ["string", "null"] }
        price: { type: number, minimum: 0 }
        year: { type: ["integer", "null"] }
        mileage: { type: ["integer", "null"] }
        engineSize: { type: ["integer", "null"] }
        image: { type: string, minLength: 1 }
        description: { type: string, minLength: 10 }
        location: { type: string, minLength: 2 }
        sellerId: { type: integer }
        premium: { type: ["boolean", "null"] }
        stock: { type: ["integer", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        workingHours: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        fuelType: { type: ["string", "null"] }
        optionals: { type: ["string", "null"] }
        tradeInfo: { type: ["string", "null"] }
        extras: { type: ["string", "null"] }
    UpdateItem:
      type: object
      properties:
        title: { type: string }
        price: { type: number }
        description: { type: string }
        status: { type: string, enum: [active, sold, pending] }
        premium: { type: boolean }
        stock: { type: integer }
        category: { type: string }
...
    OrderItem:
      type: object
      required: [id, orderId, itemId, title, image, price, qty]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        itemId: { type: integer }
        title: { type: string }
        image: { type: string }
        price: { type: number }
        qty: { type: integer }
    Order:
      type: object
      required: [id, buyerId, sellerId, status, subtotal, discount, total, paymentMethod, shippingAddress, createdAt, items]
      properties:
        id: { type: integer }
        buyerId: { type: integer }
        sellerId: { type: integer }
        status: { type: string, enum: [pending, paid, shipped, delivered, cancelled] }
        subtotal: { type: number }
        discount: { type: number }
        total: { type: number }
        couponCode: { type: ["string", "null"] }
        paymentMethod: { type: string, enum: [pix, card, boleto] }
        shippingAddress: { type: string }
        createdAt: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
    CartLine:
      type: object
      required: [itemId, qty]
      properties:
        itemId: { type: integer }
        qty: { type: integer }
    CreateOrderInput:
      type: object
      required: [buyerId, paymentMethod, shippingAddress, lines]
      properties:
        buyerId: { type: integer }
        paymentMethod: { type: string, enum: [pix, card, boleto] }
        shippingAddress: { type: string }
        couponCode: { type: ["string", "null"] }
        lines:
          type: array
          items: { $ref: "#/components/schemas/CartLine" }
    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, paid, shipped, delivered, cancelled] }
    Review:
      type: object
      required: [id, itemId, userId, rating, comment, createdAt, userName]
      properties:
        id: { type: integer }
        itemId: { type: integer }
        userId: { type: integer }
        rating: { type: integer }
        comment: { type: string }
        createdAt: { type: string }
        userName: { type: string }
    CreateReviewInput:
      type: object
      required: [itemId, userId, rating]
      properties:
        itemId: { type: integer }
        userId: { type: integer }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
    Coupon:
      type: object
      required: [id, code, type, value, minOrder, active, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        type: { type: string, enum: [percent, fixed] }
        value: { type: number }
        minOrder: { type: number }
        active: { type: boolean }
        createdAt: { type: string }
    CreateCouponInput:
      type: object
      required: [code, type, value]
      properties:
        code: { type: string }
        type: { type: string, enum: [percent, fixed] }
        value: { type: number }
        minOrder: { type: ["number", "null"] }
        active: { type: ["boolean", "null"] }
    ValidateCouponInput:
      type: object
      required: [code, subtotal]
      properties:
        code: { type: string }
        subtotal: { type: number }
    ValidateCouponResult:
      type: object
      required: [valid, discount]
      properties:
        valid: { type: boolean }
        discount: { type: number }
        message: { type: ["string", "null"] }
        coupon: { $ref: "#/components/schemas/Coupon" }
...