# lib/api-zod/src/generated/types/createReviewInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MotoHub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewInput {
  itemId: number;
  userId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
}