# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MotoHub API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  buyerId: number;
  sellerId: number;
  status: OrderStatus;
  subtotal: number;
  discount: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
  paymentMethod: OrderPaymentMethod;
  shippingAddress: string;
  createdAt: string;
  items: OrderItem[];
}